make client install more idempotent

+++ cookbooks/chef/recipes/client.rb
@@ -16,10 +16,13 @@ end
 
 execute "Register client node with chef server" do
   command "#{node[:chef][:client_path]} -t #{`cat /etc/chef/validation_token`}"
+  only_if { File.exists? "/etc/chef/validation_token" }
+  not_if { File.exists?("/var/chef/cache/registration") }
 end
 
 execute "Remove the validation token" do
   command "rm /etc/chef/validation_token"
+  only_if { File.exists? "/etc/chef/validation_token" }
 end
 
 runit_service "chef-client"
